+++ zcash_primitives/src/merkle_tree.rs
@@ -1,14 +1,13 @@
 //! Implementation of a Merkle tree of commitments used to prove the existence of notes.
 
+use bridgetree::Frontier;
 use byteorder::{LittleEndian, ReadBytesExt};
-use incrementalmerkletree::Hashable;
+use incrementalmerkletree::{Hashable, Level};
 use std::collections::VecDeque;
 use std::convert::TryInto;
 use std::io::{self, Read, Write};
 use std::iter::repeat;
 use zcash_encoding::{Optional, Vector};
-
-use crate::sapling::SAPLING_COMMITMENT_TREE_DEPTH_U8;
 
 pub mod incremental;
 
@@ -34,25 +33,22 @@ impl<Node: Hashable> PathFiller<Node> {
         }
     }
 
-    fn next(&mut self, depth: u8) -> Node {
+    fn next(&mut self, level: Level) -> Node {
         self.queue
             .pop_front()
-            .unwrap_or_else(|| Node::empty_root(depth.into()))
+            .unwrap_or_else(|| Node::empty_root(level))
     }
 }
 
 /// A Merkle tree of note commitments.
-///
-/// The depth of the Merkle tree is fixed at 32, equal to the depth of the Sapling
-/// commitment tree.
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CommitmentTree<Node> {
+pub struct CommitmentTree<Node, const DEPTH: u8> {
     pub(crate) left: Option<Node>,
     pub(crate) right: Option<Node>,
     pub(crate) parents: Vec<Option<Node>>,
 }
 
-impl<Node> CommitmentTree<Node> {
+impl<Node, const DEPTH: u8> CommitmentTree<Node, DEPTH> {
     /// Creates an empty tree.
     pub fn empty() -> Self {
         CommitmentTree {
@@ -62,7 +58,7 @@ impl<Node> CommitmentTree<Node> {
         }
     }
 
-    pub fn from_frontier<const DEPTH: u8>(frontier: &bridgetree::Frontier<Node, DEPTH>) -> Self
+    pub fn from_frontier(frontier: &Frontier<Node, DEPTH>) -> Self
     where
         Node: Clone,
     {
@@ -97,12 +93,12 @@ impl<Node> CommitmentTree<Node> {
         })
     }
 
-    pub fn to_frontier<const DEPTH: u8>(&self) -> bridgetree::Frontier<Node, DEPTH>
+    pub fn to_frontier(&self) -> Frontier<Node, DEPTH>
     where
         Node: Hashable + Clone,
     {
         if self.size() == 0 {
-            bridgetree::Frontier::empty()
+            Frontier::empty()
         } else {
             let ommers_iter = self.parents.iter().filter_map(|v| v.as_ref()).cloned();
             let (leaf, ommers) = match (self.left.as_ref(), self.right.as_ref()) {
@@ -116,7 +112,7 @@ impl<Node> CommitmentTree<Node> {
 
             // If a frontier cannot be successfully constructed from the
             // parts of a commitment tree, it is a programming error.
-            bridgetree::Frontier::from_parts((self.size() - 1).into(), leaf, ommers)
+            Frontier::from_parts((self.size() - 1).into(), leaf, ommers)
                 .expect("Frontier should be constructable from CommitmentTree.")
         }
     }
@@ -154,9 +150,12 @@ impl<Node> CommitmentTree<Node> {
     }
 }
 
-impl<Node: Hashable + HashSer> CommitmentTree<Node> {
+impl<Node: Hashable + Clone, const DEPTH: u8> CommitmentTree<Node, DEPTH> {
     /// Reads a `CommitmentTree` from its serialized form.
-    pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
+    pub fn read<R: Read>(mut reader: R) -> io::Result<Self>
+    where
+        Node: HashSer,
+    {
         let left = Optional::read(&mut reader, Node::read)?;
         let right = Optional::read(&mut reader, Node::read)?;
         let parents = Vector::read(&mut reader, |r| Optional::read(r, Node::read))?;
@@ -169,23 +168,26 @@ impl<Node: Hashable + HashSer> CommitmentTree<Node> {
     }
 
     /// Serializes this tree as an array of bytes.
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
+    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()>
+    where
+        Node: HashSer,
+    {
         Optional::write(&mut writer, self.left.as_ref(), |w, n| n.write(w))?;
         Optional::write(&mut writer, self.right.as_ref(), |w, n| n.write(w))?;
         Vector::write(&mut writer, &self.parents, |w, e| {
             Optional::write(w, e.as_ref(), |w, n| n.write(w))
         })
     }
-}
 
-impl<Node: Hashable + Clone> CommitmentTree<Node> {
     /// Adds a leaf node to the tree.
     ///
     /// Returns an error if the tree is full.
+    #[allow(clippy::result_unit_err)]
     pub fn append(&mut self, node: Node) -> Result<(), ()> {
-        self.append_inner(node, SAPLING_COMMITMENT_TREE_DEPTH_U8)
+        self.append_inner(node, DEPTH)
     }
 
+    #[allow(clippy::result_unit_err)]
     fn append_inner(&mut self, node: Node, depth: u8) -> Result<(), ()> {
         if self.is_complete(depth) {
             // Tree is full
@@ -223,7 +225,7 @@ impl<Node: Hashable + Clone> CommitmentTree<Node> {
 
     /// Returns the current root of the tree.
     pub fn root(&self) -> Node {
-        self.root_inner(SAPLING_COMMITMENT_TREE_DEPTH_U8, PathFiller::empty())
+        self.root_inner(DEPTH, PathFiller::empty())
     }
 
     fn root_inner(&self, depth: u8, mut filler: PathFiller<Node>) -> Node {
@@ -231,16 +233,17 @@ impl<Node: Hashable + Clone> CommitmentTree<Node> {
 
         // 1) Hash left and right leaves together.
         //    - Empty leaves are used as needed.
+        //    - Note that `filler.next` is side-effecting and so cannot be factored out.
         let leaf_root = Node::combine(
             0.into(),
             &self
                 .left
                 .as_ref()
-                .map_or_else(|| filler.next(0), |n| n.clone()),
+                .map_or_else(|| filler.next(0.into()), |n| n.clone()),
             &self
                 .right
                 .as_ref()
-                .map_or_else(|| filler.next(0), |n| n.clone()),
+                .map_or_else(|| filler.next(0.into()), |n| n.clone()),
         );
 
         // 2) Extend the parents to the desired depth with None values, then hash from leaf to
@@ -251,10 +254,10 @@ impl<Node: Hashable + Clone> CommitmentTree<Node> {
             .take((depth - 1).into())
             .zip(0u8..)
             .fold(leaf_root, |root, (p, i)| {
-                let parent_level = i + 1;
+                let level = Level::from(i + 1);
                 match p {
-                    Some(node) => Node::combine(parent_level.into(), node, &root),
-                    None => Node::combine(parent_level.into(), &root, &filler.next(parent_level)),
+                    Some(node) => Node::combine(level, node, &root),
+                    None => Node::combine(level, &root, &filler.next(level)),
                 }
             })
     }
@@ -269,54 +272,53 @@ impl<Node: Hashable + Clone> CommitmentTree<Node> {
 /// # Examples
 ///
 /// ```
-/// use ff::{Field, PrimeField};
-/// use rand_core::OsRng;
-/// use zcash_primitives::{
-///     merkle_tree::{CommitmentTree, IncrementalWitness},
-///     sapling::Node,
+/// use zcash_primitives::merkle_tree::{
+///   CommitmentTree,
+///   IncrementalWitness,
+///   testing::TestNode
 /// };
 ///
-/// let mut rng = OsRng;
+/// let mut tree = CommitmentTree::<TestNode, 8>::empty();
 ///
-/// let mut tree = CommitmentTree::<Node>::empty();
-///
-/// tree.append(Node::from_scalar(bls12_381::Scalar::random(&mut rng)));
-/// tree.append(Node::from_scalar(bls12_381::Scalar::random(&mut rng)));
-/// let mut witness = IncrementalWitness::from_tree(&tree);
+/// tree.append(TestNode(0));
+/// tree.append(TestNode(1));
+/// let mut witness = IncrementalWitness::from_tree(tree.clone());
 /// assert_eq!(witness.position(), 1);
 /// assert_eq!(tree.root(), witness.root());
 ///
-/// let cmu = Node::from_scalar(bls12_381::Scalar::random(&mut rng));
-/// tree.append(cmu);
-/// witness.append(cmu);
+/// let next = TestNode(2);
+/// tree.append(next.clone());
+/// witness.append(next);
 /// assert_eq!(tree.root(), witness.root());
 /// ```
 #[derive(Clone, Debug)]
-pub struct IncrementalWitness<Node> {
-    tree: CommitmentTree<Node>,
+pub struct IncrementalWitness<Node, const DEPTH: u8> {
+    tree: CommitmentTree<Node, DEPTH>,
     filled: Vec<Node>,
     cursor_depth: u8,
-    cursor: Option<CommitmentTree<Node>>,
+    cursor: Option<CommitmentTree<Node, DEPTH>>,
 }
 
-impl<Node: Hashable + HashSer + Clone> IncrementalWitness<Node> {
+impl<Node, const DEPTH: u8> IncrementalWitness<Node, DEPTH> {
     /// Creates an `IncrementalWitness` for the most recent commitment added to the given
     /// [`CommitmentTree`].
-    pub fn from_tree(tree: &CommitmentTree<Node>) -> IncrementalWitness<Node> {
+    pub fn from_tree(tree: CommitmentTree<Node, DEPTH>) -> IncrementalWitness<Node, DEPTH> {
         IncrementalWitness {
-            tree: tree.clone(),
+            tree,
             filled: vec![],
             cursor_depth: 0,
             cursor: None,
         }
     }
+}
 
+impl<Node: Hashable + HashSer + Clone, const DEPTH: u8> IncrementalWitness<Node, DEPTH> {
     /// Reads an `IncrementalWitness` from its serialized form.
     #[allow(clippy::redundant_closure)]
-    pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
-        let tree = CommitmentTree::read(&mut reader)?;
+    pub fn read<R: Read>(mut reader: R) -> io::Result<IncrementalWitness<Node, DEPTH>> {
+        let tree = CommitmentTree::<Node, DEPTH>::read(&mut reader)?;
         let filled = Vector::read(&mut reader, |r| Node::read(r))?;
-        let cursor = Optional::read(&mut reader, CommitmentTree::read)?;
+        let cursor = Optional::read(&mut reader, CommitmentTree::<Node, DEPTH>::read)?;
 
         let mut witness = IncrementalWitness {
             tree,
@@ -395,10 +397,12 @@ impl<Node: Hashable + HashSer + Clone> IncrementalWitness<Node> {
     /// Tracks a leaf node that has been added to the underlying tree.
     ///
     /// Returns an error if the tree is full.
+    #[allow(clippy::result_unit_err)]
     pub fn append(&mut self, node: Node) -> Result<(), ()> {
-        self.append_inner(node, SAPLING_COMMITMENT_TREE_DEPTH_U8)
+        self.append_inner(node, DEPTH)
     }
 
+    #[allow(clippy::result_unit_err)]
     fn append_inner(&mut self, node: Node, depth: u8) -> Result<(), ()> {
         if let Some(mut cursor) = self.cursor.take() {
             cursor
@@ -433,7 +437,7 @@ impl<Node: Hashable + HashSer + Clone> IncrementalWitness<Node> {
 
     /// Returns the current root of the tree corresponding to the witness.
     pub fn root(&self) -> Node {
-        self.root_inner(SAPLING_COMMITMENT_TREE_DEPTH_U8)
+        self.root_inner(DEPTH)
     }
 
     fn root_inner(&self, depth: u8) -> Node {
@@ -441,11 +445,11 @@ impl<Node: Hashable + HashSer + Clone> IncrementalWitness<Node> {
     }
 
     /// Returns the current witness, or None if the tree is empty.
-    pub fn path(&self) -> Option<MerklePath<Node>> {
-        self.path_inner(SAPLING_COMMITMENT_TREE_DEPTH_U8)
+    pub fn path(&self) -> Option<MerklePath<Node, DEPTH>> {
+        self.path_inner(DEPTH)
     }
 
-    fn path_inner(&self, depth: u8) -> Option<MerklePath<Node>> {
+    fn path_inner(&self, depth: u8) -> Option<MerklePath<Node, DEPTH>> {
         let mut filler = self.filler();
         let mut auth_path = Vec::new();
 
@@ -453,7 +457,7 @@ impl<Node: Hashable + HashSer + Clone> IncrementalWitness<Node> {
             if self.tree.right.is_some() {
                 auth_path.push((node.clone(), true));
             } else {
-                auth_path.push((filler.next(0), false));
+                auth_path.push((filler.next(0.into()), false));
             }
         } else {
             // Can't create an authentication path for the beginning of the tree
@@ -470,41 +474,51 @@ impl<Node: Hashable + HashSer + Clone> IncrementalWitness<Node> {
         {
             auth_path.push(match p {
                 Some(node) => (node.clone(), true),
-                None => (filler.next(i + 1), false),
+                None => (filler.next(Level::from(i + 1)), false),
             });
         }
 
-        assert_eq!(auth_path.len(), depth.into());
+        assert_eq!(auth_path.len(), usize::from(depth));
 
-        Some(MerklePath::from_path(auth_path, self.position() as u64))
+        MerklePath::from_path(auth_path, self.position() as u64).ok()
     }
 }
 
 /// A path from a position in a particular commitment tree to the root of that tree.
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct MerklePath<Node: Hashable> {
-    pub auth_path: Vec<(Node, bool)>,
-    pub position: u64,
+pub struct MerklePath<Node, const DEPTH: u8> {
+    auth_path: Vec<(Node, bool)>,
+    position: u64,
 }
 
-impl<Node: Hashable + HashSer> MerklePath<Node> {
+impl<Node, const DEPTH: u8> MerklePath<Node, DEPTH> {
+    pub fn auth_path(&self) -> &[(Node, bool)] {
+        &self.auth_path
+    }
+
+    pub fn position(&self) -> u64 {
+        self.position
+    }
+
     /// Constructs a Merkle path directly from a path and position.
-    pub fn from_path(auth_path: Vec<(Node, bool)>, position: u64) -> Self {
-        MerklePath {
-            auth_path,
-            position,
+    pub fn from_path(auth_path: Vec<(Node, bool)>, position: u64) -> Result<Self, ()> {
+        if auth_path.len() == usize::from(DEPTH) {
+            Ok(MerklePath {
+                auth_path,
+                position,
+            })
+        } else {
+            Err(())
         }
     }
+}
 
+impl<Node: Hashable + HashSer, const DEPTH: u8> MerklePath<Node, DEPTH> {
     /// Reads a Merkle path from its serialized form.
-    pub fn from_slice(witness: &[u8]) -> Result<Self, ()> {
-        Self::from_slice_with_depth(witness, SAPLING_COMMITMENT_TREE_DEPTH_U8)
-    }
-
-    fn from_slice_with_depth(mut witness: &[u8], depth: u8) -> Result<Self, ()> {
-        // Skip the first byte, which should be "depth" to signify the length of
+    pub fn from_slice(mut witness: &[u8]) -> Result<Self, ()> {
+        // Skip the first byte, which should be DEPTH to signify the length of
         // the following vector of Pedersen hashes.
-        if witness[0] != depth {
+        if witness[0] != DEPTH {
             return Err(());
         }
         witness = &witness[1..];
@@ -533,7 +547,7 @@ impl<Node: Hashable + HashSer> MerklePath<Node> {
                 }
             })
             .collect::<Result<Vec<_>, _>>()?;
-        if auth_path.len() != depth.into() {
+        if auth_path.len() != usize::from(DEPTH) {
             return Err(());
         }
 
@@ -583,9 +597,8 @@ mod tests {
     use incrementalmerkletree::Hashable;
     use proptest::prelude::*;
     use proptest::strategy::Strategy;
-    use std::io::{self, Read, Write};
 
-    use crate::sapling::{testing::arb_node, Node};
+    use crate::sapling::{self, testing::arb_node, Node};
 
     use super::{
         testing::{arb_commitment_tree, TestNode},
@@ -628,68 +641,6 @@ mod tests {
         "fbc2f4300c01f0b7820d00e3347c8da4ee614674376cbc45359daa54f9b5493e",
     ];
 
-    const TESTING_DEPTH: u8 = 4;
-
-    #[derive(Debug)]
-    struct TestCommitmentTree(CommitmentTree<Node>);
-
-    impl TestCommitmentTree {
-        fn new() -> Self {
-            TestCommitmentTree(CommitmentTree::empty())
-        }
-
-        pub fn read<R: Read>(reader: R) -> io::Result<Self> {
-            let tree = CommitmentTree::read(reader)?;
-            Ok(TestCommitmentTree(tree))
-        }
-
-        pub fn write<W: Write>(&self, writer: W) -> io::Result<()> {
-            self.0.write(writer)
-        }
-
-        fn size(&self) -> usize {
-            self.0.size()
-        }
-
-        fn append(&mut self, node: Node) -> Result<(), ()> {
-            self.0.append_inner(node, TESTING_DEPTH)
-        }
-
-        fn root(&self) -> Node {
-            self.0.root_inner(TESTING_DEPTH, PathFiller::empty())
-        }
-    }
-
-    #[derive(Debug)]
-    struct TestIncrementalWitness(IncrementalWitness<Node>);
-
-    impl TestIncrementalWitness {
-        fn from_tree(tree: &TestCommitmentTree) -> Self {
-            TestIncrementalWitness(IncrementalWitness::from_tree(&tree.0))
-        }
-
-        pub fn read<R: Read>(reader: R) -> io::Result<Self> {
-            let witness = IncrementalWitness::read(reader)?;
-            Ok(TestIncrementalWitness(witness))
-        }
-
-        pub fn write<W: Write>(&self, writer: W) -> io::Result<()> {
-            self.0.write(writer)
-        }
-
-        fn append(&mut self, node: Node) -> Result<(), ()> {
-            self.0.append_inner(node, TESTING_DEPTH)
-        }
-
-        fn root(&self) -> Node {
-            self.0.root_inner(TESTING_DEPTH)
-        }
-
-        fn path(&self) -> Option<MerklePath<Node>> {
-            self.0.path_inner(TESTING_DEPTH)
-        }
-    }
-
     #[test]
     fn empty_root_test_vectors() {
         let mut tmp = [0u8; 32];
@@ -704,7 +655,7 @@ mod tests {
     #[test]
     fn sapling_empty_root() {
         let mut tmp = [0u8; 32];
-        CommitmentTree::<Node>::empty()
+        sapling::CommitmentTree::empty()
             .root()
             .write(&mut tmp[..])
             .expect("length is 32 bytes");
@@ -716,7 +667,7 @@ mod tests {
 
     #[test]
     fn empty_commitment_tree_roots() {
-        let tree = CommitmentTree::<Node>::empty();
+        let tree = sapling::CommitmentTree::empty();
         let mut tmp = [0u8; 32];
         for (&expected, i) in HEX_EMPTY_ROOTS.iter().zip(0u8..).skip(1) {
             tree.root_inner(i, PathFiller::empty())
@@ -1053,40 +1004,46 @@ mod tests {
             "01f43e3aac61e5a753062d4d0508c26ceaf5e4c0c58ba3c956e104b5d2cf67c41c0003015991131c5c25911b35fcea2a8343e2dfd7a4d5b45493390e0cb184394d91c34901002df68503da9247dfde6585cb8c9fa94897cf21735f8fc1b32116ef474de05c010d6b42350c11df4fcc17987c13d8492ba4e8b3f31eb9baff9be5d8890cfa512d013a3661bc12b72646c94bc6c92796e81953985ee62d80a9ec3645a9a95740ac1500",
         ];
 
+        const TESTING_DEPTH: u8 = 4;
+
         fn assert_root_eq(root: Node, expected: &str) {
             let mut tmp = [0u8; 32];
             root.write(&mut tmp[..]).expect("length is 32 bytes");
             assert_eq!(hex::encode(tmp), expected);
         }
 
-        fn assert_tree_ser_eq(tree: &TestCommitmentTree, expected: &str) {
+        fn assert_tree_ser_eq(tree: &CommitmentTree<Node, TESTING_DEPTH>, expected: &str) {
             // Check that the tree matches its encoding
             let mut tmp = Vec::new();
             tree.write(&mut tmp).unwrap();
             assert_eq!(hex::encode(&tmp[..]), expected);
 
             // Check round-trip encoding
-            let decoded = TestCommitmentTree::read(&hex::decode(expected).unwrap()[..]).unwrap();
+            let decoded: CommitmentTree<Node, TESTING_DEPTH> =
+                CommitmentTree::read(&hex::decode(expected).unwrap()[..]).unwrap();
             tmp.clear();
             decoded.write(&mut tmp).unwrap();
             assert_eq!(hex::encode(tmp), expected);
         }
 
-        fn assert_witness_ser_eq(witness: &TestIncrementalWitness, expected: &str) {
+        fn assert_witness_ser_eq(
+            witness: &IncrementalWitness<Node, TESTING_DEPTH>,
+            expected: &str,
+        ) {
             // Check that the witness matches its encoding
             let mut tmp = Vec::new();
             witness.write(&mut tmp).unwrap();
             assert_eq!(hex::encode(&tmp[..]), expected);
 
             // Check round-trip encoding
-            let decoded =
-                TestIncrementalWitness::read(&hex::decode(expected).unwrap()[..]).unwrap();
+            let decoded: IncrementalWitness<Node, TESTING_DEPTH> =
+                IncrementalWitness::read(&hex::decode(expected).unwrap()[..]).unwrap();
             tmp.clear();
             decoded.write(&mut tmp).unwrap();
             assert_eq!(hex::encode(tmp), expected);
         }
 
-        let mut tree = TestCommitmentTree::new();
+        let mut tree = CommitmentTree::<Node, TESTING_DEPTH>::empty();
         assert_eq!(tree.size(), 0);
 
         let mut witnesses = vec![];
@@ -1099,7 +1056,7 @@ mod tests {
             let cmu = Node::new(cmu[..].try_into().unwrap());
 
             // Witness here
-            witnesses.push((TestIncrementalWitness::from_tree(&tree), last_cmu));
+            witnesses.push((IncrementalWitness::from_tree(tree.clone()), last_cmu));
 
             // Now append a commitment to the tree
             assert!(tree.append(cmu).is_ok());
@@ -1119,11 +1076,8 @@ mod tests {
 
                 if let Some(leaf) = leaf {
                     let path = witness.path().expect("should be able to create a path");
-                    let expected = MerklePath::from_slice_with_depth(
-                        &hex::decode(paths[paths_i]).unwrap(),
-                        TESTING_DEPTH,
-                    )
-                    .unwrap();
+                    let expected =
+                        MerklePath::from_slice(&hex::decode(paths[paths_i]).unwrap()).unwrap();
                     assert_eq!(path, expected);
 
                     assert_eq!(path.root(*leaf), witness.root());
@@ -1176,7 +1130,7 @@ mod tests {
 
     proptest! {
         #[test]
-        fn prop_commitment_tree_roundtrip_str(ct in arb_commitment_tree(32, any::<char>().prop_map(|c| c.to_string()), 8)) {
+        fn prop_commitment_tree_roundtrip_str(ct in arb_commitment_tree::<_, _, 8>(32, any::<char>().prop_map(|c| c.to_string()))) {
             let frontier: Frontier<String, 8> = ct.to_frontier();
             let ct0 = CommitmentTree::from_frontier(&frontier);
             assert_eq!(ct, ct0);
@@ -1185,7 +1139,7 @@ mod tests {
         }
 
         #[test]
-        fn prop_commitment_tree_roundtrip_node(ct in arb_commitment_tree(32, arb_node(), 8)) {
+        fn prop_commitment_tree_roundtrip_node(ct in arb_commitment_tree::<_, _, 8>(32, arb_node())) {
             let frontier: Frontier<Node, 8> = ct.to_frontier();
             let ct0 = CommitmentTree::from_frontier(&frontier);
             assert_eq!(ct, ct0);
@@ -1194,21 +1148,21 @@ mod tests {
         }
 
         #[test]
-        fn prop_commitment_tree_roundtrip_ser(ct in arb_commitment_tree(32, arb_node(), 8)) {
+        fn prop_commitment_tree_roundtrip_ser(ct in arb_commitment_tree::<_, _, 8>(32, arb_node())) {
             let mut serialized = vec![];
             assert_matches!(ct.write(&mut serialized), Ok(()));
-            assert_matches!(CommitmentTree::read(&serialized[..]), Ok(ct_out) if ct == ct_out);
+            assert_matches!(CommitmentTree::<_, 8>::read(&serialized[..]), Ok(ct_out) if ct == ct_out);
         }
     }
 
     #[test]
     fn test_commitment_tree_complete() {
-        let mut t: CommitmentTree<TestNode> = CommitmentTree::empty();
+        let mut t: CommitmentTree<TestNode, 32> = CommitmentTree::empty();
         for n in 1u64..=32 {
             t.append(TestNode(n)).unwrap();
             // every tree of a power-of-two height is complete
             let is_complete = n.count_ones() == 1;
-            let level = 63 - n.leading_zeros(); //log2
+            let level = usize::BITS - 1 - n.leading_zeros(); //log2
             assert_eq!(
                 is_complete,
                 t.is_complete(level.try_into().unwrap()),
@@ -1238,18 +1192,21 @@ pub mod testing {
 
     use super::{CommitmentTree, HashSer};
 
-    pub fn arb_commitment_tree<Node: Hashable + Clone + Debug, T: Strategy<Value = Node>>(
+    pub fn arb_commitment_tree<
+        Node: Hashable + Clone + Debug,
+        T: Strategy<Value = Node>,
+        const DEPTH: u8,
+    >(
         min_size: usize,
         arb_node: T,
-        depth: u8,
-    ) -> impl Strategy<Value = CommitmentTree<Node>> {
-        assert!((1 << depth) >= min_size + 100);
+    ) -> impl Strategy<Value = CommitmentTree<Node, DEPTH>> {
+        assert!((1 << DEPTH) >= min_size + 100);
         vec(arb_node, min_size..(min_size + 100)).prop_map(move |v| {
             let mut tree = CommitmentTree::empty();
             for node in v.into_iter() {
                 tree.append(node).unwrap();
             }
-            tree.parents.resize_with((depth - 1).into(), || None);
+            tree.parents.resize_with((DEPTH - 1).into(), || None);
             tree
         })
     }